+++ gui/extension/CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Additions
 
-- 
+-
 
 ### Fixes
 
-- 
+- BUG#37919552 Rest Routine with Async Task does not return a result in the TypeScript SDK
+-
 
 ## Changes in 1.19.9+9.3.1